move non-generic methods from A to Sdk.Create
