+++ docs/dummies.md
@@ -42,10 +42,13 @@ A.CollectionOfDummy<Book>(10);
 
 This will return an `IList` containing 10 dummy `Book` instances.
 
-For cases where the type of dummy isn't statically known, non-generic methods are also available:
+For cases where the type of dummy isn't statically known, non-generic methods are also available. These are usually only required when writing extensions for FakeItEasy, so they live in the `FakeItEasy.Sdk` namespace:
 ```csharp
-object book = A.Dummy(typeof(Book));
-IList<object> books = A.CollectionOfFake(typeof(Book), 10);
+using FakeItEasy.Sdk;
+...
+var type = GetTypeOfDummy();
+object dummy = Create.Dummy(type);
+IList<object> dummies = Create.CollectionOfDummy(type, 10);
 ```
 
 ## How FakeItEasy uses them
