+++ src/FakeItEasy/A.cs
@@ -49,27 +49,6 @@ namespace FakeItEasy
         }
 
         /// <summary>
-        /// Creates a fake object of the specified type.
-        /// </summary>
-        /// <param name="typeOfFake">The type of fake object to create.</param>
-        /// <returns>A fake object.</returns>
-        public static object Fake(Type typeOfFake)
-        {
-            return FakeCreator.CreateFake(typeOfFake, x => { });
-        }
-
-        /// <summary>
-        /// Creates a fake object of the specified type.
-        /// </summary>
-        /// <param name="typeOfFake">The type of fake object to create.</param>
-        /// <param name="optionsBuilder">A lambda where options for the built fake object can be specified.</param>
-        /// <returns>A fake object.</returns>
-        public static object Fake(Type typeOfFake, Action<IFakeOptions> optionsBuilder)
-        {
-            return FakeCreator.CreateFake(typeOfFake, optionsBuilder);
-        }
-
-        /// <summary>
         /// Creates a collection of fakes of the specified type.
         /// </summary>
         /// <typeparam name="T">The type of fakes to create.</typeparam>
@@ -95,29 +74,6 @@ namespace FakeItEasy
         }
 
         /// <summary>
-        /// Creates a collection of fakes of the specified type.
-        /// </summary>
-        /// <param name="typeOfFake">The type of fakes to create.</param>
-        /// <param name="numberOfFakes">The number of fakes in the collection.</param>
-        /// <returns>A collection of fake objects of the specified type.</returns>
-        public static IList<object> CollectionOfFake(Type typeOfFake, int numberOfFakes)
-        {
-            return FakeCreator.CollectionOfFake(typeOfFake, numberOfFakes);
-        }
-
-        /// <summary>
-        /// Creates a collection of fakes of the specified type.
-        /// </summary>
-        /// <param name="typeOfFake">The type of fakes to create.</param>
-        /// <param name="numberOfFakes">The number of fakes in the collection.</param>
-        /// <param name="optionsBuilder">A lambda where options for the built fake object can be specified.</param>
-        /// <returns>A collection of fake objects of the specified type.</returns>
-        public static IList<object> CollectionOfFake(Type typeOfFake, int numberOfFakes, Action<IFakeOptions> optionsBuilder)
-        {
-            return FakeCreator.CollectionOfFake(typeOfFake, numberOfFakes, optionsBuilder);
-        }
-
-        /// <summary>
         /// Gets a dummy object of the specified type. The value of a dummy object
         /// should be irrelevant. Dummy objects should not be configured.
         /// </summary>
@@ -132,19 +88,6 @@ namespace FakeItEasy
         }
 
         /// <summary>
-        /// Gets a dummy object of the specified type. The value of a dummy object
-        /// should be irrelevant. Dummy objects should not be configured.
-        /// </summary>
-        /// <param name="typeOfDummy">The type of dummy to return.</param>
-        /// <returns>A dummy object of the specified type.</returns>
-        /// <exception cref="ArgumentException">Dummies of the specified type can not be created.</exception>
-        [EditorBrowsable(EditorBrowsableState.Advanced)]
-        public static object Dummy(Type typeOfDummy)
-        {
-            return FakeCreator.CreateDummy(typeOfDummy);
-        }
-
-        /// <summary>
         /// Creates a collection of dummies of the specified type.
         /// </summary>
         /// <typeparam name="T">The type of dummies to create.</typeparam>
@@ -156,19 +99,6 @@ namespace FakeItEasy
         public static IList<T> CollectionOfDummy<T>(int numberOfDummies)
         {
             return FakeCreator.CollectionOfDummy<T>(numberOfDummies);
-        }
-
-        /// <summary>
-        /// Creates a collection of dummies of the specified type.
-        /// </summary>
-        /// <param name="typeOfDummy">The type of dummy to return.</param>
-        /// <param name="numberOfDummies">The number of dummies in the collection.</param>
-        /// <returns>A collection of dummy objects of the specified type.</returns>
-        /// <exception cref="ArgumentException">Dummies of the specified type can not be created.</exception>
-        [EditorBrowsable(EditorBrowsableState.Advanced)]
-        public static IList<object> CollectionOfDummy(Type typeOfDummy, int numberOfDummies)
-        {
-            return FakeCreator.CollectionOfDummy(typeOfDummy, numberOfDummies);
         }
 
         /// <summary>
